// src/models/reportType.ts
export enum ReportType {
    SPEED_TRAP = 0,
    ACCIDENT = 1,
    TRAFFIC_JAM = 2,
    STATIONARY_VEHICLE = 3,
    ROADWORKS = 4,
    OBSTACLE = 5,
    SLIPPERY_ROAD = 6,
    BUMPY_ROAD = 7,
    GHOST_DRIVER = 9,
    MAX_SPEED = 10,
    OTHER = 11,
    SPEED_CAM = 12,
    ANIMAL_ON_ROAD = 19,
    DANGEROUS_SITUATION = 34,
    PERSON_ON_ROAD = 36,
    TRAFFIC_CAM = 39,
    CONTROL = 40,
}